MYSQL fully working on local

+++ philz-customer-v2/src/main/java/com/example/IService.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 public interface IService<T> {
 	Collection<T> findAll();
 	
-	Optional<T> findById(Long id);
+	Optional<T> findById(Integer id);
 	
 	T saveOrUpdate(T t);
 	
-	String deleteById(Long id);
+	String deleteById(Integer id);
 }
